norminette

+++ utils/ft_putnbr_ptr.c
@@ -6,16 +6,16 @@
 /*   By: caucane <christian.aucane@learner.42.tech  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/17 16:52:35 by caucane           #+#    #+#             */
-/*   Updated: 2025/11/17 16:52:51 by caucane          ###   ########.fr       */
+/*   Updated: 2025/11/17 19:08:47 by caucane          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "utils.h"
 
-void	ft_putnbr_ptr(unsigned long nb, int *chars_count)
+void	ft_putnbr_ptr(uintptr_t nb, int *chars_count)
 {
-	unsigned long	base_len;
-	char			*base;
+	uintptr_t	base_len;
+	char		*base;
 
 	base = "0123456789abcdef";
 	base_len = 16;
